check and update to the norm

+++ libft/ft_strncmp.c
@@ -6,13 +6,13 @@
 /*   By: bebrandt <benoit.brandt@proton.me>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/16 11:08:53 by bebrandt          #+#    #+#             */
-/*   Updated: 2023/10/17 12:27:57 by bebrandt         ###   ########.fr       */
+/*   Updated: 2023/10/17 16:37:40 by bebrandt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strncmp(const char *s1, const char *s2, long unsigned int n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	while (*s1 && *s2 && *s1 == *s2 && (n - 1))
 	{
